refactor: 모달 버튼 공통으로 사용할 수 있도록 네이밍 변경,  옵셔널 속성으로 변경
